début de l'implémentation du contrôleur pour l'envoie de fichier

+++ trunk/ChatSystemG2/chatSystem/controller/ChatController.java
@@ -92,7 +92,7 @@ public class ChatController extends Controller implements GuiToCont, NiToCont {
     }
 
     @Override
-    public void performSendFileRequest(String name, int size, String idRemoteSystem) {
+    public void performSendFileRequest(String name, long size, String idRemoteSystem) {
        System.out.println("Send file request to be send to " + idRemoteSystem + ", modifying the model");
        this.fileTransferts.addTransfert(name, size, idRemoteSystem, this.transfertID);
        this.transfertID ++;
