+++ trunk/ChatSystemG2/chatSystem/model/FileTransferts.java
@@ -27,7 +27,7 @@ public class FileTransferts extends Model{
      * @param idRemoteSystem sender's id
      * @param idTransfert id given to the transfer
      */
-    public void addTransfert(String name, int size, String idRemoteSystem, int idTransfert) {
+    public void addTransfert(String name, long size, String idRemoteSystem, int idTransfert) {
             this.FileModel.add(new FileTransfertInformation(size,idRemoteSystem, name, idTransfert));
     }
